requirements.txt
publicsuffix2
requests
ruamel.yaml
validators
